refactor(step3) : Product 패키지 분리
