+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain;
 
+import kitchenpos.product.domain.Product;
+
 import javax.persistence.*;
 import java.util.UUID;
