+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,7 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
 import kitchenpos.domain.*;
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.profanity.domain.PurgomalumClient;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
