+++ src/main/java/kitchenpos/profanity/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.profanity.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
