+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,10 +1,13 @@
-package kitchenpos.application;
+package kitchenpos.product.application;
 
+import kitchenpos.profanity.infrastructure.FakePurgomalumClient;
+import kitchenpos.application.InMemoryMenuRepository;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuRepository;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRepository;
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.product.infrastructure.InMemoryProductRepository;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.profanity.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
