+++ src/test/java/kitchenpos/profanity/infrastructure/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
-package kitchenpos.application;
+package kitchenpos.profanity.infrastructure;
 
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.profanity.domain.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
